split out docker tests, added test for multiple docker images.

+++ ci_utils.sh
@@ -17,14 +17,18 @@ iso8601_datetime() {
 }
 
 ci_log() {
-    local logdir=".log"
+    local logdir="${SUBMODULES_PATH}/.log"
     local logfile="${logdir}/${project}.ci.log"
     mkdir -p "${logdir}"
     touch "${logfile}"
+    if [[ ! -f "${logfile}" ]]; then
+        echo "ERROR FNF"
+        exit 1
+    fi
     printf "\n" | tee -a "${logfile}"
     printf "CI Task Started:\n" | tee -a "${logfile}"
     printf " ISO8601 Datetime: %s\n" "$(iso8601_datetime)" | tee -a "${logfile}"
-    #IFS='\n'
+    IFS='\n'
     while read -r input; do
         echo "$input" 2>&1 | tee -a >( sed -r 's/\x1b\[[0-9;]*m//g' >> "${logfile}") 
     done
